ast: rename `Expr::LitExpr` to `Expr::Lit` and add `BlockOrigin` to crate
