+++ compiler/hash-ast/src/lib.rs
@@ -4,6 +4,7 @@
 
 pub mod ast;
 pub mod node_map;
+pub mod origin;
 pub mod tree;
 pub mod visitor {
     pub use super::ast::{
