+++ compiler/hash-untyped-semantics/src/analysis/mod.rs
@@ -6,7 +6,7 @@ pub(crate) mod params;
 mod pat;
 
 use crossbeam_channel::Sender;
-use hash_ast::ast::AstNodeRef;
+use hash_ast::{ast::AstNodeRef, origin::BlockOrigin};
 use hash_reporting::diagnostic::Diagnostics;
 use hash_source::{
     location::{SourceLocation, Span},
@@ -15,7 +15,6 @@ use hash_source::{
 
 use crate::diagnostics::{
     error::{AnalysisError, AnalysisErrorKind},
-    origins::BlockOrigin,
     warning::{AnalysisWarning, AnalysisWarningKind},
     AnalyserDiagnostics, AnalysisDiagnostic,
 };
